rebase

+++ src/KOKKOS/atom_vec_full_kokkos.h
@@ -50,7 +50,7 @@ class AtomVecFullKokkos : public AtomVecKokkos {
   int pack_data_hybrid(int, double *);
   void write_data(FILE *, int, double **);
   int write_data_hybrid(FILE *, double *);
-  bigint memory_usage();
+  double memory_usage();
 
   void grow_pointers();
   int pack_border_kokkos(int n, DAT::tdual_int_2d k_sendlist,
