+++ src/nbin.h
@@ -42,7 +42,7 @@ class NBin : protected Pointers {
   void post_constructor(class NeighRequest *);
   virtual void copy_neighbor_info();
   virtual void bin_atoms_setup(int);
-  bigint memory_usage();
+  double memory_usage();
 
   virtual void setup_bins(int) = 0;
   virtual void bin_atoms() = 0;
